+++ src/app/blog/page.tsx
@@ -1,46 +1,189 @@
-import { Suspense } from 'react'
-import { CallOut } from '@codegouvfr/react-dsfr/CallOut'
-import Section from '@/components/Section'
-import HtmlViewer from '@/components/HtmlViewer'
-import { getMarkdown } from '@/lib/markdown'
-import type { DataType } from '@/lib/markdown'
-import Breadcrumb from '@/layouts/Breadcrumb'
-import { TextWrapper } from './page.styled'
-import pageTitle from '@/utils/pageTitle'
+'use client'
 
-export const metadata = pageTitle('Blog en maintenance')
+import { useCallback, useEffect, useState, useMemo, Suspense } from 'react'
+import { useRouter } from 'next/navigation'
+import { useSearchParams } from 'next/navigation'
+import { usePathname } from 'next/navigation'
+import { Pagination } from '@codegouvfr/react-dsfr/Pagination'
+import Tag from '@codegouvfr/react-dsfr/Tag'
 
-export default async function Home() {
-  const { contentHtml, data }: { contentHtml?: string, data?: DataType } = await getMarkdown('blog-en-maintenance') || {}
+import BlogGrid from '@/components/BlogGrid'
+import { getPosts, getTags } from '@/lib/blog'
+
+import { TagsWrapper, TagWrapper } from './page.styled'
+import useClientSidePageTitle from '@/utils/useClientSidePageTitle'
+
+interface PostMeta {
+  pagination?: {
+    page: number
+    limit: number
+    pages: number
+    total: number
+  }
+}
+
+function BlogView() {
+  useClientSidePageTitle('Blog')
+  const router = useRouter()
+  const searchParams = useSearchParams()
+  const pathname = usePathname()
+  const [allTags, setAllTags] = useState(new Set())
+  const [highlightedPosts, setHighlightedPosts] = useState([])
+  const [posts, setPosts] = useState([])
+  const [meta, setMeta] = useState<PostMeta>({})
+  const [loadedQueryKey, setLoadedQueryKey] = useState('')
+  const [intialLoading, setIntialLoading] = useState(true)
+
+  const nbHighlightedPosts = 3
+  const nbPost = 15
+  const page: number = useMemo(() => Number(searchParams?.get('page') || 1), [searchParams])
+  const tags: Set<string> = useMemo(() => new Set(decodeURI(searchParams?.get('tags') || '')?.split(',').filter(t => t).sort() || []), [searchParams])
+  const queryKey = useMemo(() => `${page}|${[...tags].join(',')}`, [page, tags])
+
+  const getPageLink = useCallback(({ page, tags: _tags }: { page?: number, tags?: string[] }) => {
+    const urlSearchParams = searchParams?.entries() as unknown as string[][]
+    const linkSearchParams = new URLSearchParams(urlSearchParams)
+    if (_tags) {
+      if (_tags.length) {
+        linkSearchParams.set('tags', _tags.filter(t => t).join(','))
+      }
+      else {
+        linkSearchParams.delete('tags')
+      }
+    }
+
+    if (page) {
+      linkSearchParams.set('page', page.toString())
+    }
+    const link = `${pathname}?${linkSearchParams.toString()}`
+    return link
+  }, [searchParams, pathname])
+
+  useEffect(() => {
+    getPosts({
+      limit: nbHighlightedPosts,
+    }, { cache: 'no-store' }).then((data) => {
+      setHighlightedPosts(data.posts)
+    })
+  }, [])
+
+  useEffect(() => {
+    let isCurrent = true
+
+    Promise.all([
+      getPosts({
+        page,
+        limit: nbPost,
+        tags: tags.size ? [...tags].join(',') : undefined,
+      }, { cache: 'no-store' }),
+      getTags({ cache: 'no-store' }),
+    ])
+      .then(([postsData, tagsData]) => {
+        if (!isCurrent) {
+          return
+        }
+
+        setPosts(postsData.posts)
+        setMeta(postsData.meta)
+        setAllTags(new Set(tagsData))
+        setLoadedQueryKey(queryKey)
+      })
+      .finally(() => {
+        if (!isCurrent) {
+          return
+        }
+
+        setIntialLoading(false)
+      })
+
+    return () => {
+      isCurrent = false
+    }
+  }, [page, tags, queryKey])
 
   return (
     <>
-      <Breadcrumb
-        currentPageLabel="Blog"
-        segments={[]}
+      {/* <Suspense fallback={<div>Loading...</div>}> */}
+      <BlogGrid
+        title="Le blog : articles et témoignages"
+        posts={highlightedPosts}
+        nbPost={3}
+        isVisible={page === 1 && !tags.size}
       />
-      <Section>
-        <TextWrapper>
-          <Suspense fallback={<p>Chargement...</p>}>
-            <article>
-              {contentHtml && <HtmlViewer html={contentHtml} />}
-            </article>
+      {/* </Suspense> */}
+
+      {/* <Suspense fallback={<div>Loading...</div>}> */}
+      <BlogGrid
+        title="Tous les articles"
+        header={(
+          <TagsWrapper>
+            {allTags.size > 1 && (
+              <TagWrapper>
+                <Tag
+                  nativeButtonProps={{
+                    onClick: () => {
+                      if (tags.size) {
+                        const pageLink = getPageLink({ page: 1, tags: [] })
+                        router.push(pageLink, { scroll: false })
+                      }
+                    },
+                  }}
+                  pressed={Boolean(!tags.size)}
+                >
+                  Tous les articles
+                </Tag>
+              </TagWrapper>
+            )}
             {
-              data?.aside && (
-                <aside>{data?.aside?.map(
-                  ({ data }) =>
-                    data?.contentHtml && (
-                      <CallOut key={`${data?.data?.title}`}>
-                        <HtmlViewer html={data?.contentHtml} />
-                      </CallOut>
-                    )
-                )}
-                </aside>
-              )
+              (Array.from(allTags) as { name: string, slug: string }[])
+                .map(({ name, slug }) => (
+                  <TagWrapper key={name}>
+                    <Tag
+                      nativeButtonProps={{
+                        onClick: () => {
+                          const newTags = new Set(tags.has(slug)
+                            ? [...tags].filter(tag => tag !== slug)
+                            : [...tags, slug])
+                          const pageLink = getPageLink({ page: 1, tags: [...newTags] })
+                          router.push(pageLink, { scroll: true })
+                        },
+                      }}
+                      pressed={tags.has(slug)}
+                    >{name}
+                    </Tag>
+                  </TagWrapper>
+                ))
             }
-          </Suspense>
-        </TextWrapper>
-      </Section>
+          </TagsWrapper>
+        )}
+        footer={(
+          <>
+            {meta.pagination && (
+              <Pagination
+                count={meta.pagination?.pages}
+                defaultPage={meta.pagination?.page}
+                getPageLinkProps={pageNumber => ({ href: getPageLink({ page: pageNumber }) })}
+                showFirstLast
+              />
+            )}
+          </>
+        )}
+        posts={
+          posts && posts.length
+            ? (posts.slice((!page || page === 1) && tags.size === 0 ? 3 : 0, nbPost))
+            : undefined
+        }
+        isLoading={!intialLoading && loadedQueryKey !== queryKey}
+      />
+      {/* </Suspense> */}
     </>
+  )
+}
+
+export default function BlogPage() {
+  return (
+    <Suspense fallback={<div>Chargement...</div>}>
+      <BlogView />
+    </Suspense>
   )
 }
